split modules tests
